refactor(ReceiverLink): move addCredits to ReceiverLink
