+++ lib/frames/flow_frame.js
@@ -144,8 +144,7 @@ var debug = require('debug')('amqp10:framing:flow'),
  * @constructor
  */
 function FlowFrame(options) {
-  FlowFrame.super_.call(this);
-  this.channel = 0;
+  FlowFrame.super_.call(this, options.channel);
   if (options instanceof DescribedType) {
     this.readPerformative(options);
     return;
