+++ lib/link.js
@@ -119,39 +119,6 @@ Link.prototype.detach = function() {
   return detachPromise;
 };
 
-Link.prototype.addCredits = function(credits, flowOptions) {
-  if (this.role === constants.linkRole.sender) {
-    throw new errors.InvalidStateError('Cannot add link credits as a sender');
-  }
-
-  var opts = flowOptions || {};
-  this.linkCredit += credits;
-  this.totalCredits += credits;
-  opts.linkCredit = this.totalCredits;
-  this.session._sessionParams.incomingWindow += credits;
-  opts.nextIncomingId = this.session._sessionParams.nextIncomingId;
-  opts.incomingWindow = this.session._sessionParams.incomingWindow;
-  opts.nextOutgoingId = this.session._sessionParams.nextOutgoingId;
-  opts.outgoingWindow = this.session._sessionParams.outgoingWindow;
-  opts.handle = this.handle;
-  opts.available = this.available;
-  opts.deliveryCount = this.deliveryCount;
-  opts.drain = false;
-  var flow = new FlowFrame(opts);
-  flow.channel = this.session.channel;
-  this.session.connection.sendFrame(flow);
-
-  var self = this;
-  return new Promise(function(resolve, reject) {
-    var onError = function(err) { reject(err); };
-    self.once(Link.ErrorReceived, onError);
-    self.once(Link.CreditChange, function() {
-      self.removeListener(Link.ErrorReceived, onError);
-      resolve();
-    });
-  });
-};
-
 Link.prototype.accept = function(message) {
   var range = u.dispositionRange(message);
   this._sendDisposition(_.defaults(range, {
